% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hydrograph.R
\name{as_hydrograph}
\alias{as_hydrograph}
\alias{as_hydrograph.data.frame}
\title{Coerce an object to an event hydrograph}
\usage{
as_hydrograph(x, ...)

\method{as_hydrograph}{data.frame}(x, times_from, flows_from, unit = 1)
}
\arguments{
\item{x}{A data frame to coerce to a hydrograph.}

\item{...}{Used for extensions to specific coercion methods.}

\item{times_from}{Column name containing timings of the hydrograph
change points, including the terminus times. Unquoted.}

\item{flows_from}{Column name containing the discharge values
corresponding to the timings in the \code{times_from} column. Unquoted.}

\item{unit}{Conversion factor for the time values; single numeric.
Useful if, for example, the timing is specified as a proportion of
the total event time.}
}
\value{
An event hydrograph, as in \code{hydrograph()}, made up of
linearly interpolated discharge-timing pairs.
}
\description{
Turn an object into an event hydrograph.
}
\examples{
df <- data.frame(when = 0:3, flows = c(2, 5, 4, 2))
h <- as_hydrograph(df, when, flows)
plot(h)
}
