% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dataframe_to_tibble.R
\name{convert_dataframe_to_tibble}
\alias{convert_dataframe_to_tibble}
\title{Possibly convert a data.frame to a tibble}
\usage{
convert_dataframe_to_tibble(res)
}
\arguments{
\item{res}{Data frame to possibly convert to tibble.}
}
\value{
Returns \code{res} if the tibble package is not installed,
and \code{as_tibble(res)} if the tibble package is installed.
}
\description{
If a user has the tibble package installed, convert a data frame
to a tibble.
}
