% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_hydrograph.R
\name{discretize_hydrograph}
\alias{discretize_hydrograph}
\title{Discretize a Hydrograph}
\usage{
discretize_hydrograph(hydrograph, niter, times_to = "time", flows_to = "flow")
}
\arguments{
\item{hydrograph}{A hydrograph object.}

\item{niter}{Number of time points in the grid.}

\item{times_to, flows_to}{Column names for the time and flow values.}
}
\value{
A data frame representation of a hydrograph.
}
\description{
Evaluates a hydrograph over a time grid.
}
\examples{
hyd_snow(5, 2) |>
  discretize_hydrograph(100) |>
  plot()
}
