% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erode.R
\name{erode}
\alias{erode}
\title{Erode a channel from GBEM}
\usage{
erode(object)
}
\arguments{
\item{object}{The output of \code{gbem()}, containing the channel change.}
}
\value{
The original cross section supplied to gbem, with possibly
modified width based on the gbem run.
}
\description{
While \code{gbem()} calculates fluvial geomorphic change, \code{erode()}
executes the change to create a modified channel or cross section.
}
