% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eroding_flow.R
\name{eroding_flow}
\alias{eroding_flow}
\title{Flow Threshold for Channel Erosion}
\usage{
eroding_flow(sx, resistance = c("ferguson", "manning"))
}
\arguments{
\item{sx}{Channel cross sections containing features needed by the
erosion engine, such as created by the \code{sx()} functions such as
\code{sx_ferguson()} or \code{sx_manning()}.
If running the engine \code{"manning-volume"}, then \code{sx} needs to have
a column named \code{xs2d} containing the two-dimensional cross section
objects extended sufficiently far beyond the banks to accommodate
erosion.}
}
\value{
Vector of discharges for each cross section.
}
\description{
For a given channel, calculates the largest flow a channel can accommodate
without eroding.
}
\examples{
library(sxchan)
sxc <- xt_sxc(c(12, 15, 18))
x <- sx_manning(sxc, grad = 0.02, d50 = 65, d84 = 100, roughness = 0.01)
eroding_flow(x, resistance = "manning")
}
