% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mu.R
\name{est_mu}
\alias{est_mu}
\title{Estimate mu value}
\usage{
est_mu(H, d)
}
\arguments{
\item{H}{Bank height ??? (meters???); numeric vector.}

\item{d}{Water depth ??? (meters???); numeric vector.}
}
\value{
Vector of \code{mu} values.
}
\description{
Estimate mu value
}
\examples{
est_mu(5, 2)

}
