% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph-snowrain.R
\name{hydrograph_rain}
\alias{hydrograph_rain}
\alias{hydrograph_snow}
\title{Snowmelt and Rainfall-Related Event Hydrographs}
\usage{
hydrograph_rain(peak, baseflow, duration = 3 * 24)

hydrograph_snow(peak, baseflow, duration = 9 * 24)
}
\arguments{
\item{peak, baseflow}{Peak and baseflow discharges; single numerics.}

\item{duration}{Length of the event. Defaults to 3 days for
rainfall-related events, and 9 days for snowmelt-related events
(in hours).}
}
\description{
Constructs event hydrographs characteristic of a rainfall or
snowmelt-related event.
}
\details{
Rainfall-related event hydrographs are triangular, rising
from baseflow to peak in 1/3 the duration, followed by a recession
to baseflow.

Snowmelt-related event hydrographs are like the rainfall hydrographs,
but remain at the peak discharge until half of the event duration.
}
\examples{
hydrograph_rain(5, 2) |> plot()
hydrograph_snow(5, 2) |> plot()

}
