% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ferguson_n.R
\name{ferguson_n}
\alias{ferguson_n}
\title{Function to use the Ferguson equation to calculate an equivalent n value}
\usage{
ferguson_n(d, S, D84)
}
\arguments{
\item{d}{Depth}

\item{S}{Energy gradient of the stream channel}

\item{D84}{84th percentile of the surface grain size distribution (mm)}
}
\value{
n value
}
\description{
Function to use the Ferguson equation to calculate an equivalent n value
}
