% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_d_crit.R
\name{find_d_crit}
\alias{find_d_crit}
\title{Calculates the critical threshold for channel widening}
\usage{
find_d_crit(Q, H, t_c84, S)
}
\arguments{
\item{Q}{Discharge carried by the stream, single numeric}

\item{H}{Effective rooting depth for vegetation, single numeric}

\item{S}{Energy gradient of the stream channel, single numeric}
}
\value{
Critical threshold for channel widening
}
\description{
Calculates the critical threshold for channel widening
}
\examples{
gbem:::find_d_crit(100, 0.35, 0.5, 0.01)

}
