% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_q_b.R
\name{find_q_b}
\alias{find_q_b}
\title{Determine if channel will widen and calculate transp, widening}
\usage{
find_q_b(d, n, d50, S)
}
\arguments{
\item{d}{Flow depth (m)}

\item{n}{Manning's n value for the main channel}

\item{d50}{50th percentile of the grain size distribution (mm)}

\item{S}{Energy gradient of the stream channel}
}
\description{
Determine if channel will widen and calculate transp, widening
}
