% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem.R
\name{gbem}
\alias{gbem}
\title{Gravel-Bed River Bank Erosion Model}
\usage{
gbem(
  hydrograph,
  sx,
  niter = 1000,
  engine = c("manning-original", "manning-volume", "ferguson"),
  side = c("both", "left", "right"),
  keep_movie = FALSE
)
}
\arguments{
\item{hydrograph}{Event hydrograph carried by the stream, with time units
in hours. See \code{hydrograph()}.}

\item{sx}{Channel cross sections containing features needed by the
erosion engine, such as created by the \code{sx()} functions such as
\code{sx_ferguson()} or \code{sx_manning()}.
If running the engine \code{"manning-volume"}, then \code{sx} needs to have
a column named \code{xs2d} containing the two-dimensional cross section
objects extended sufficiently far beyond the banks to accommodate
erosion.}

\item{niter}{Number of time chunks to discretize the hydrograph into.}

\item{engine}{Paradigm for modelling channel resistance; one of
\code{"ferguson"} for using Ferguson's equation,
\code{"manning-original"} for Manning's equation (original algorithm on
simplified channels), or
\code{"manning-volume"} for Manning's equation on 2D cross sections.}

\item{side}{Which side of the channel to erode? One of
\code{"both"} (erode equally on both sides), \code{"left"}, or \code{"right"}
(erode entirely on one side of the channel).}

\item{keep_movie}{Logical; keep all intermediate points? Only applies
to the constrained \code{const} outputs and for \verb{"manning-volume}" engine.
I'll probably remove this argument later because the output size can
become massive, but for now it's good for model exploration.}
}
\value{
A list of the following components:
\itemize{
\item \code{sx}: The original cross section.
\item \code{dw_max}: maximum widening.
\item \code{dw_const}: change in width constrained by transport capacity, the most
\item \code{v_b}: transport capacity * time. Volume of transport that can be moved
by the river. Matrix, with rows representing cross sections corresponding
to the rows of \code{sx}; columns are time steps.
\item \code{resistance}: the resistance paradigm used.
}
}
\description{
Runs an event hydrograph through a channel and applies an erosion engine
to determine lateral erosion.
}
\details{
This algorithm implements one of two paradigms for modelling
channel resistance.
\enumerate{
\item \strong{Manning's Equation}: FILL_THIS_IN
\item \strong{Ferguson's Equation}: FILL_THIS_IN
}

An error is thrown if the input \code{sx} channel does not have sufficiently
described properties.
}
\examples{
library(sf)
library(sxchan)

## Set up the channel.
cross_sections <- xt_generate_sxc(demo_bankline, n = 20)
n <- length(cross_sections)
channel <- sx_manning(
  cross_sections, grad = 0.02, d50 = 45, d84 = 90,
  roughness = append(rep(0.01, n / 2), rep(0.05, n / 2))
)

## Create an event hydrograph.
hg <- hyd_snow(200, baseflow = 20)

## Run the hydrograph through the channel, using Manning's method.
demo_gbem <- gbem(hg, channel, niter = 100, engine = "manning-original")

## Erode the channel
(new_channel <- erode(demo_gbem)) # Erosion occurs

plot(demo_bankline)
plot(st_geometry(new_channel), add = TRUE, col = "blue")

## Run a smaller event through that does no erosion.
q <- min(eroding_flow(channel, engine = "manning-original"))
hg2 <- hyd_rain(q / 2, baseflow = q / 10)
demo_gbem2 <- gbem(hg2, channel, niter = 100, engine = "manning-original")
new_channel2 <- erode(demo_gbem2)

## No erosion:
identical(st_geometry(channel), st_geometry(new_channel2))
}
