% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem0.R
\name{gbem0}
\alias{gbem0}
\title{Single iteration of the GBEM algorithm}
\usage{
gbem0(flow, duration, cross_section)
}
\arguments{
\item{flow}{Discharge carried by the stream.}

\item{duration}{Time for which flow acts on the stream channel (hrs).}

\item{cross_section}{A \code{"cross_section"} object.}
}
\value{
A list of the following components:
\itemize{
\item \code{dw_pred}: predicted widening.
\item \code{dw_const}: change in width constrained by transport capacity, the most
\item \code{v_b}: transport capacity * time. Volume of transport that can be moved
by the river.
\item \code{cross_section}: The original cross section.
}
}
\description{
Run the gravel-bed river bank erosion model to determine channel
changes for a constant flow over a small duration.
}
\seealso{
\code{\link[=erode]{erode()}}
}
