% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem0_manning.R
\name{gbem0_manning}
\alias{gbem0_manning}
\title{Single iteration of the GBEM algorithm}
\usage{
gbem0_manning(flow, duration, width, grad, d50, d84, roughness, rootdepth)
}
\arguments{
\item{flow}{Discharge carried by the stream.}

\item{duration}{Time for which flow acts on the stream channel (hrs).}

\item{width}{Cross section width}

\item{grad}{Channel gradient at the cross section.}

\item{d50, d84}{Grain size distribution quantities (50th and 84th percentile).}

\item{roughness}{Roughness value.}

\item{rootdepth}{Effective rooting depth; defaults to 0.}
}
\value{
A list of the following components:
\itemize{
\item \code{dw_pred}: predicted widening.
\item \code{dw_const}: change in width constrained by transport capacity, the most
\item \code{v_b}: transport capacity * time. Volume of transport that can be moved
by the river.
}
}
\description{
Run the gravel-bed river bank erosion model to determine channel
changes for a constant flow over a small duration.
}
\seealso{
\code{\link[=erode]{erode()}}
}
