% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_hydrograph.R
\name{gbem2}
\alias{gbem2}
\title{Uses gbem to estimate the channel change during a flood hydrograph
with Q values for every hour of the floods}
\usage{
gbem2(hydrograph, cross_section, niter = 100)
}
\arguments{
\item{hydrograph}{Event hydrograph carried by the stream, with time units
in hours.}

\item{cross_section}{Cross section of the stream.}

\item{niter}{Number of iterations to run the algorithm over.}
}
\value{
A discretized event hydrograph with erosion widths and volumes
at each time step.
}
\description{
Uses gbem to estimate the channel change during a flood hydrograph
with Q values for every hour of the floods
}
\examples{
hg <- hydrograph_snow(5, 2)
cs <- cross_section(3, grad = 0.01, d50 = 0.1, d84 = 0.5, roughness = 0.01)
g <- gbem2(hg, cs)
erode(g)
}
