% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem_ch.R
\name{gbem_ch}
\alias{gbem_ch}
\title{GBEM, but on a channel.}
\usage{
gbem_ch(hydrograph, channel, niter = 1000)
}
\arguments{
\item{hydrograph}{Event hydrograph carried by the stream, with time units
in hours.}

\item{niter}{Number of iterations to run the algorithm over.}
}
\value{
Channel object, from sxchan.
}
\description{
GBEM, but on a channel.
}
