% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_hydrograph.R
\name{generic_hydrograph}
\alias{generic_hydrograph}
\title{Uses gbem to estimate the channel change during a flood hydrograph
with Q values for every hour of the floods}
\usage{
generic_hydrograph(Q, n, d84, d50, W, S, H = 0)
}
\arguments{
\item{Q}{Discharge carried by the stream}

\item{n}{Manning's n value for the main channel}

\item{d84}{84th percentile of the surface grain size distribution (mm)}

\item{d50}{50th percentile of the grain size distribution (mm)}

\item{W}{Water surface width at the beginning of time interval T}

\item{S}{Energy gradient of the stream channel}

\item{H}{Effective rooting depth for vegetation}
}
\description{
Uses gbem to estimate the channel change during a flood hydrograph
with Q values for every hour of the floods
}
