% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hydrograph.R
\name{is_hydrograph}
\alias{is_hydrograph}
\title{Test if an object is a hydrograph}
\usage{
is_hydrograph(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if an object inherits the class \code{"hydrograph"}.
}
\description{
Returns \code{TRUE} if an object inherits the class \code{"hydrograph"}, and
\code{FALSE} otherwise.
}
