% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_stable_channel.R
\name{min_stable_channel}
\alias{min_stable_channel}
\title{Minimum-Width Stable Channel}
\usage{
min_stable_channel(sx, flow, resistance = c("ferguson", "manning"))
}
\arguments{
\item{sx}{Channel cross sections containing features needed by the
erosion engine, such as created by the \code{sx()} functions such as
\code{sx_ferguson()} or \code{sx_manning()}.
If running the engine \code{"manning-volume"}, then \code{sx} needs to have
a column named \code{xs2d} containing the two-dimensional cross section
objects extended sufficiently far beyond the banks to accommodate
erosion.}

\item{flow}{Discharge carried by the stream; single numeric.}
}
\value{
The original \code{sx} channel object, with cross section widths
modified to match the minimum width required to accommodate the flow
without eroding.
}
\description{
Calculates the channel with the smallest width that remains stable
under a given flow.
}
\examples{
library(sxchan)
x <- sx_manning(
  xt_sxc(c(1, 1, 1)), grad = 0.01, d50 = 45, d84 = 80,
  roughness = c(0.01, 0.02, 0.03)
)
y <- min_stable_channel(x, flow = 5, resistance = "manning")

xt_width(x)
xt_width(y)
}
