% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_channel.R
\name{new_channel}
\alias{new_channel}
\title{Constructor function for channel objects}
\usage{
new_channel(l, ..., class = character())
}
\arguments{
\item{l}{List containing the components of a channel object.}

\item{...}{Attributes to add to the object.}

\item{class}{If making a subclass, specify its name here.}
}
\value{
An object of class \code{"channel"}, although not necessarily with
valid properties.
}
\description{
Constructor function for channel objects
}
