% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gbem.R
\name{new_gbem}
\alias{new_gbem}
\title{Create a gbem object}
\usage{
new_gbem(x, ..., class = character())
}
\arguments{
\item{x}{Object to add gbem class to.}

\item{...}{Other attributes to add to the object.}

\item{class}{If specifying a subclass, indicate it here.}
}
\value{
Object with class "gbem" and attributes specified in \code{...}.
}
\description{
Internal function for making an object of class "gbem".
}
