% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph.R
\name{new_hydrograph}
\alias{new_hydrograph}
\title{Constructor function for (event) hydrograph objects}
\usage{
new_hydrograph(fun, times, peak, ..., class = character())
}
\arguments{
\item{fun}{Function of time that represents the hydrograph}

\item{times}{Vector of length 2 indicating the start and end time of the
event hydrograph (numeric).}

\item{peak}{Magnitude of the peak discharge in the event hydrograph;
single numeric.}
}
\value{
Object of class "hydrograph".
}
\description{
Constructor function for (event) hydrograph objects
}
