% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{peak}
\alias{peak}
\title{Hydrograph peak}
\usage{
peak(hydrograph)
}
\arguments{
\item{hydrograph}{Object of class "hydrograph".}
}
\value{
The peak discharge of the event hydrograph; single numeric.
}
\description{
Returns the peak of an event hydrograph.
}
