% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_hydrograph.R
\name{rain_hydrograph}
\alias{rain_hydrograph}
\title{Uses GBEM to estimate the channel change during a simulated flood
hydrograph due to a rain dominated flood}
\usage{
rain_hydrograph(Q, baseflow, n, d84, d50, W, S, H = 0, raw = FALSE)
}
\arguments{
\item{Q}{Discharge carried by the stream}

\item{baseflow}{Mean annual flow}

\item{n}{Manning's n value for the main channel}

\item{d84}{84th percentile of the surface grain size distribution (mm)}

\item{d50}{50th percentile of the grain size distribution (mm)}

\item{W}{Water surface width at the beginning of time interval T}

\item{S}{Energy gradient of the stream channel}

\item{H}{Effective rooting depth for vegetation}
}
\value{
Parameters of the flood event to create a hydrograph with.
}
\description{
Uses GBEM to estimate the channel change during a simulated flood
hydrograph due to a rain dominated flood
}
