% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_hydrograph_old.R, R/snow_hydrograph_old.R
\name{rain_hydrograph_old}
\alias{rain_hydrograph_old}
\alias{snow_hydrograph_old}
\title{Uses GBEM to estimate the channel change during a simulated flood
hydrograph due to a snow or rainfall dominated flood}
\usage{
rain_hydrograph_old(Q, baseflow, n, d84, d50, W, S, H = 0, raw = FALSE)

snow_hydrograph_old(Q, baseflow, n, d84, d50, W, S, H = 0, raw = FALSE)
}
\arguments{
\item{Q}{Discharge carried by the stream}

\item{baseflow}{Mean annual flow}

\item{n}{Manning's n value for the main channel}

\item{d84}{84th percentile of the surface grain size distribution (mm)}

\item{d50}{50th percentile of the grain size distribution (mm)}

\item{W}{Water surface width at the beginning of time interval T}

\item{S}{Energy gradient of the stream channel}

\item{H}{Effective rooting depth for vegetation}

\item{raw}{Logical; return the raw data from the analysis?}
}
\value{
Parameters of the flood event to create a hydrograph with
}
\description{
Uses GBEM to estimate the channel change during a simulated flood
hydrograph due to a snow or rainfall dominated flood
}
