% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.hydrograph.R
\name{range.hydrograph}
\alias{range.hydrograph}
\title{Event hydrograph timing}
\usage{
\method{range}{hydrograph}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Event hydrograph (only one allowed).}

\item{na.rm}{Not used.}
}
\value{
Vector of length 2 indicating the start and end times of the
event hydrograph
}
\description{
Returns the time frame for which the event hydrograph occurs.
}
