% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sx.R
\name{sx_manning}
\alias{sx_manning}
\alias{sx_ferguson}
\alias{sx}
\title{Add features to cross sections}
\usage{
sx_manning(object, ..., grad, d50, d84, roughness, rootdepth = 0)

sx_ferguson(object, ..., grad, d50, roughness, rootdepth = 0)

sx(object, ...)
}
\arguments{
\item{object}{An object of class "sxc" from the sxchan package.}

\item{...}{Other features to include as columns in the data frame. Since
this is passed to \code{sf::st_sf()}, not all column names are permitted;
see details.}

\item{grad}{Channel gradient at the cross section.}

\item{d50, d84}{Grain size distribution quantities (50th and 84th percentile).}

\item{roughness}{Roughness value.}

\item{rootdepth}{Effective rooting depth; defaults to 0.}
}
\value{
An object of class "sf", with the geometry column being an "sxc"
object.
}
\description{
Cross sections from the sxchan package are a collection of line segments.
This function stacks these line segments into a data frame column,
so that features of the cross sections can be specified in other columns.
Entries are recycled to the number of cross sections.
}
\details{
For downstream modification of columns, use your favourite method for
modifying data frames.

Examples of effective rooting depth for vegetation, \code{rootdepth},
are:
\itemize{
\item grassy banks, no trees / shrubs: \code{rootdepth = 0.35}.
\item 1 to 5\% tree / shrub cover: \code{rootdepth = 0.50}.
\item 5 to 50\% tree / shrub cover: \code{rootdepth = 0.90}.
\item more than 50\% tree / shrub cover: \code{rootdepth = 1.10}.
}
}
\note{
The \code{sx()} function is a simple wrapper around \code{sf::st_sf()},
aside from some basic checks to try to be helpful.
It exists to support the pairing of specialized features required
for different erosion engines, like \code{sx_manning()}. Otherwise,
\code{sx()} may not be needed, with \code{sf::st_sf()} sufficing.
}
\examples{
library(sxchan)
x <- xt_sxc(1:5)
sx_manning(
  x, grad = 0.1, d50 = 1, d84 = 2, swimmability = 4,
  roughness = c(0.01, 0.02, 0.01, 0.01, 0.1), rootdepth = 0
)
}
