% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_channel.R
\name{validate_channel}
\alias{validate_channel}
\title{Validator function for channel objects}
\usage{
validate_channel(channel)
}
\arguments{
\item{channel}{Object of class \code{"channel"}.}
}
\value{
Returns the original \code{channel} object. Note that this function is
intended to be run for its side effects: namely, throwing an error if the
channel is invalid.
}
\description{
Checks that an object of class \code{"channel"} has a valid structure and valid
channel properties.
}
