% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph.R
\name{validate_hydrograph}
\alias{validate_hydrograph}
\title{Validator function for hydrograph objects}
\usage{
validate_hydrograph(hydrograph)
}
\arguments{
\item{hydrograph}{A hydrograph object.}
}
\description{
Validator function for hydrograph objects
}
