% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_sx.R
\name{validate_sx}
\alias{validate_sx}
\title{Validator function for sx objects}
\usage{
validate_sx(sx, resistance = NULL)
}
\arguments{
\item{sx}{Object of class \code{"sx"}.}

\item{resistance}{Optional specification of a resistance paradigm to
check against; see \code{gbem()} for options. Keep \code{NULL} to only check
the overall structure of the input \code{sx} object.}
}
\value{
Returns the original \code{sx} object, invisibly.
}
\description{
Checks that an object is a valid \code{sx} object. When an erosion engine is
provided, additional checks are made to ensure the provided features are
sufficient for running the erosion engine. This function is run for its
side effects, throwing an error if invalid.
}
