% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_width.R
\name{ch_width}
\alias{ch_width}
\title{Extract Channel Width}
\usage{
ch_width(cross_section)
}
\arguments{
\item{cross_section}{A \code{"cross_section"} object representing a stream's
cross section.}
}
\value{
Cross section width; single numeric.
}
\description{
Extract Channel Width
}
