% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph-snowrain.R
\name{hyd_snow}
\alias{hyd_snow}
\alias{hyd_rain}
\alias{hyd_const}
\title{Pre-set Event Hydrographs}
\usage{
hyd_snow(peak, baseflow, duration = 3 * 24)

hyd_rain(peak, baseflow, duration = 9 * 24)

hyd_const(peak, duration = 1)
}
\arguments{
\item{peak, baseflow}{Peak and baseflow discharges; single numerics.}

\item{duration}{Length of the event. Defaults to 3 days for
rainfall-related events, and 9 days for snowmelt-related events
(in hours).}
}
\description{
Constructs event hydrographs with characteristic shapes:
}
\details{
\itemize{
\item \code{hyd_rain()} is characteristic of a rainfall-related event.
\item \code{hyd_snow()} is characteristic of a snowmelt-related event.
\item \code{hyd_const()} is a flat hydrograph.
}

Rainfall-related event hydrographs are triangular, rising
from baseflow to peak in 1/3 the duration, followed by a recession
to baseflow.

Snowmelt-related event hydrographs are like the rainfall hydrographs,
but remain at the peak discharge until half of the event duration.
}
\examples{
hyd_rain(5, 2) |> plot()
hyd_snow(5, 2) |> plot()
hyd_const(5) |> plot()

}
