% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_d_crit.R
\name{find_d_crit}
\alias{find_d_crit}
\title{Calculates the critical threshold for channel widening}
\usage{
find_d_crit(H, t_c84, S)
}
\arguments{
\item{H}{Effective rooting depth for vegetation; single numeric.}

\item{t_c84}{FILL_THIS_IN}

\item{S}{Energy gradient of the stream channel; single numeric.}
}
\value{
Critical threshold for channel widening.
}
\description{
FILL_THIS_IN: THRESHOLD OF WHAT?
}
\examples{
gbem:::find_d_crit(0.35, 0.5, 0.01)
}
