% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem.R
\name{gbem}
\alias{gbem}
\title{Gravel-Bed River Bank Erosion Model}
\usage{
gbem(hydrograph, sx, niter = 1000, resistance = c("ferguson", "manning"))
}
\arguments{
\item{hydrograph}{Event hydrograph carried by the stream, with time units
in hours. See \code{hydrograph()}.}

\item{sx}{Channel cross sections containing features needed by the
erosion engine, such as created by the \code{sx()} functions such as
\code{sx_ferguson()} or \code{sx_manning()}.}

\item{niter}{Number of time chunks to discretize the hydrograph into.}

\item{resistance}{Paradigm for modelling channel resistance; one of
\code{"ferguson"} for using Ferguson's equation, or \code{"manning"} for Manning's
equation.}
}
\value{
A list of the following components:
\itemize{
\item \code{dw_pred}: predicted widening.
\item \code{dw_const}: change in width constrained by transport capacity, the most
\item \code{v_b}: transport capacity * time. Volume of transport that can be moved
by the river.
\item \code{sx}: The original cross section.
\item \code{resistance}: the resistance paradigm used.
}
}
\description{
Runs an event hydrograph through a channel and applies an erosion engine
to determine lateral erosion.
}
\details{
This algorithm implements one of two paradigms for modelling
channel resistance.
\enumerate{
\item \strong{Manning's Equation}: FILL_THIS_IN
\item \strong{Ferguson's Equation}: FILL_THIS_IN
}

An error is thrown if the input \code{sx} channel does not have sufficiently
described properties.
}
\examples{
library(sf)
library(sxchan)

## Set up the channel.
cross_sections <- xt_generate_sxc(demo_bankline, n = 10)
channel <- sx_manning(
  cross_sections, grad = 0.02, d50 = 45, d84 = 90, roughness = 0.01
)

## Create an event hydrograph.
hg <- hyd_snow(15, baseflow = 5)

## Run the hydrograph through the channel, using Manning's method.
demo_gbem <- gbem(hg, channel, niter = 100, resistance = "manning")

## Erode the channel
(new_channel <- erode(demo_gbem)) # Erosion occurs

plot(demo_bankline)
plot(st_geometry(new_channel), add = TRUE, col = "blue")

## Run the event a second time through
demo_gbem2 <- gbem(hg, new_channel, niter = 100, resistance = "manning")
new_channel2 <- erode(demo_gbem2)

## No erosion:
identical(st_geometry(new_channel), st_geometry(new_channel2))
}
