% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem.R
\name{gbem}
\alias{gbem}
\title{Gravel-Bed River Bank Erosion Model}
\usage{
gbem(hydrograph, cross_section, niter = 1000)
}
\arguments{
\item{hydrograph}{Event hydrograph carried by the stream, with time units
in hours.}

\item{cross_section}{A cross section object representing stream
cross sections; see \code{sxchan::xt_sx()}. Must have}

\item{niter}{Number of iterations to run the algorithm over.}
}
\value{
A list of the following components:
\itemize{
\item \code{dw_pred}: predicted widening.
\item \code{dw_const}: change in width constrained by transport capacity, the most
\item \code{v_b}: transport capacity * time. Volume of transport that can be moved
by the river.
\item \code{cross_section}: The original cross section.
\item \code{event}: A discretized event hydrograph with erosion widths and volumes
at each time step.
}
}
\description{
Runs the GBEM algorithm to determine erosion of a channel cross section
from a supplied event hydrograph.
}
\details{
The hydrograph is first discretized into \code{niter} constant flows,
and the GBEM algorithm is iterated on those flows.
}
\examples{
cs <- cross_section(9, grad = 0.02, d50 = 45, d84 = 90, roughness = 0.01)
hg <- hyd_snow(15, baseflow = 5)
g <- gbem(hg, cs)
erode(g) # Erosion occurs

ch_width(cs) <- 100
erode(gbem(hg, cs)) # No erosion
}
