% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem.R
\name{gbem}
\alias{gbem}
\title{Main GBEM Workhorse}
\usage{
gbem(Q, t, n, D84, D50, W, S, H = 0)
}
\arguments{
\item{Q}{discharge carried by the stream}

\item{t}{time for which Q acts on the stream channel (hrs)}

\item{name}{description}
}
\value{
description
}
\description{
Core function of gravel-bed river bank erosion model,
this is intended to become the work horse function for STOCHASIM and SGBEM
}
