% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem_iterated.R
\name{gbem}
\alias{gbem}
\title{Gravel-Bed River Bank Erosion Model}
\usage{
gbem(hydrograph, cross_section, niter = 1000)
}
\arguments{
\item{hydrograph}{Event hydrograph carried by the stream, with time units
in hours.}

\item{cross_section}{A \code{"cross_section"} object representing a stream's
cross section.}

\item{niter}{Number of iterations to run the algorithm over.}
}
\value{
A list of the following components:
\itemize{
\item \code{dw_pred}: predicted widening.
\item \code{dw_const}: change in width constrained by transport capacity, the most
\item \code{v_b}: transport capacity * time. Volume of transport that can be moved
by the river.
\item \code{cross_section}: The original cross section.
\item \code{event}: A discretized event hydrograph with erosion widths and volumes
at each time step.
}
}
\description{
Runs the GBEM algorithm to determine erosion of a channel cross section
from a supplied event hydrograph.
}
\details{
The hydrograph is first discretized into \code{niter} constant flows,
and the GBEM algorithm is iterated on those flows.
}
\examples{
hg <- hyd_snow(5, 2)
cs <- cross_section(3, grad = 0.01, d50 = 0.1, d84 = 0.5, roughness = 0.01)
g <- gbem(hg, cs)
erode(g)
}
