% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbem.R
\name{gbem}
\alias{gbem}
\title{Main GBEM Workhorse}
\usage{
gbem(Q, t, cross_section)
}
\arguments{
\item{Q}{Discharge carried by the stream}

\item{t}{Time for which Q acts on the stream channel (hrs)}

\item{cross_section}{A \code{"cross_section"} object.}
}
\value{
A list of the following components:
\itemize{
\item \code{dw_pred}: predicted widening.
\item \code{dw_const}: change in width constrained by transport capacity, the most
\item \code{v_b}: transport capacity * time. Volume of transport that can be moved
by the river.
\item \code{cross_section}: The original cross section.
}
}
\description{
Run the gravel-bed river bank erosion model to determine channel
changes.
}
\examples{
cs <- cross_section(5, grad = 0.01, d50 = 0.5, d84 = 0.5, roughness = 0.02)
gbem(100, 1, cs)
}
\seealso{
\code{\link[=erode]{erode()}}
}
