% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph.R
\name{hydrograph}
\alias{hydrograph}
\title{Construct an event hydrograph}
\usage{
hydrograph(..., unit = 1, env = rlang::caller_env())
}
\arguments{
\item{...}{Pairs of discharges and timing separated by \code{~}.}

\item{unit}{Conversion factor for the time values; single numeric.
Useful if, for example, the timing is specified as a proportion of
the total event time.}

\item{env}{Environment for which to search for bindings, in the
case that discharge-timing pairs are specified symbolically.}
}
\value{
A data frame of two columns: \code{time} into the event, and
\code{discharge} at that time. Returns a tibble if the tibble package is
installed.
}
\description{
Construct an event hydrograph by linearly connecting pairs of
discharge and timing.
}
\examples{
# Hydrograph over 10 units of time:
my_hydrograph <- hydrograph(2 ~ 0, 5 ~ 1 / 3, 2 ~ 1, unit = 10)
plot(my_hydrograph, n = 1000)

# Hydrograph spanning 24 hours:
my_hydrograph <- hydrograph(3 ~ 0, 5 ~ 10, 4.8 ~ 15, 3 ~ 24)
plot(my_hydrograph, n = 1000)
}
