% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_stable_channel.R
\name{min_stable_channel}
\alias{min_stable_channel}
\title{Minimum-Width Stable Channel}
\usage{
min_stable_channel(cross_section, flow)
}
\arguments{
\item{cross_section}{A \code{"cross_section"} object.}

\item{flow}{Discharge carried by the stream.}
}
\value{
The original cross section with width modified to be the minimum
width to accommodate the flow without eroding.
}
\description{
Calculates the channel with the smallest width that remains stable
under a given flow.
}
\examples{
cs <- cross_section(3, grad = 0.01, d50 = 45, d84 = 80, roughness = 0.01)
min_stable_channel(cs, 5)
}
